Knife now prefers DIR/config.rb to DIR/knife.rb
`knife.rb` is not deprecated at this time.

The precedence of directories is the same. Directories are searched in
the same order, so that the presence of config.rb in a low precedence
directory, e.g., `~/.chef/config.rb` does not prevent knife from loading
a knife.rb located in a higer precedence directory, e.g.,
`$PWD/.chef/knife.rb`.